+++ native/src/main/java/io/ballerina/stdlib/graphql/runtime/utils/Utils.java
@@ -40,6 +40,8 @@ public class Utils {
 
     static final String EXECUTE_RESOURCE_FUNCTION = "executeQueryResource";
 
+    static final String EXECUTE_INTERCEPTOR_FUNCTION = "executeInterceptor";
+
     // Internal type names
     public static final String ERROR_TYPE = "Error";
     public static final String CONTEXT_OBJECT = "Context";
@@ -59,6 +61,11 @@ public class Utils {
                                                                                     getModule().getName(),
                                                                                     getModule().getMajorVersion(),
                                                                                     EXECUTE_RESOURCE_FUNCTION);
+
+    public static final StrandMetadata INTERCEPTOR_EXECUTION_STRAND = new StrandMetadata(getModule().getOrg(),
+                                                                                         getModule().getName(),
+                                                                                         getModule().getMajorVersion(),
+                                                                                         EXECUTE_INTERCEPTOR_FUNCTION);
 
     public static BError createError(String message, String errorTypeName) {
         return ErrorCreator.createError(getModule(), errorTypeName, StringUtils.fromString(message), null, null);